# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel CRUD Test Assignment",
        "version": "0.1"
    },
    "paths": {
        "/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get all customers",
                "description": "Retrieve a list of all customers.",
                "operationId": "30ce4124bc07a0d013b69974ff225a5c",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Customer"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "description": "Create a new customer record.",
                "operationId": "8350a1a36d2b5522ea4158253eff28a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "firstname": [
                                                    "The firstname field is required."
                                                ],
                                                "email": [
                                                    "The email field is required.",
                                                    "The email must be a valid email address."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/customers/{id}": {
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update a customer",
                "description": "Update an existing customer record.",
                "operationId": "7f856cee51366cc0a4c5e1aa6a02547a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the customer to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Customer not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Delete a customer",
                "description": "Delete an existing customer record.",
                "operationId": "4cee47d6b826da6ed7d852b5aff46d19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the customer to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Customer deleted successfully"
                    },
                    "404": {
                        "description": "Customer not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Customer not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Customer": {
                "title": "Customer",
                "required": [
                    "firstname",
                    "lastname",
                    "date_of_birth",
                    "phone_number",
                    "email",
                    "bank_account_number"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier for the customer",
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstname": {
                        "description": "The firstname of the customer",
                        "type": "string",
                        "example": "John"
                    },
                    "lastname": {
                        "description": "The lastname of the customer",
                        "type": "string",
                        "example": "Doe"
                    },
                    "date_of_birth": {
                        "description": "The date of birth of the customer",
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "phone_number": {
                        "description": "The phone number of the customer",
                        "type": "string",
                        "example": "1234567890"
                    },
                    "email": {
                        "description": "The email address of the customer",
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "bank_account_number": {
                        "description": "The bank account number of the customer",
                        "type": "string",
                        "example": "123456789"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Customers",
            "description": "API Endpoints for Customers"
        }
    ]
}